Extract federation constants from BridgeConstants to their own classes. Create /federation package to have all federation-related classes. Move new classes to /federation/constants package.
